+++ src/Konsole/Controls/IListView.cs
@@ -1,0 +1,11 @@
+﻿namespace Konsole
+{
+    public interface IListView
+    {
+        (string name, int width)[] Columns { get; set; }
+        int selectedItemIndex { get; set; }
+
+        (string name, int width)[] GetResizedColumns();
+        void Refresh();
+    }
+}