+++ src/Konsole/Window_WriteWriteLine.cs
@@ -49,6 +49,34 @@ namespace Konsole
             }
         }
 
+        public void WriteLine(Colors colors, string text)
+        {
+            var _colors = Colors;
+            try
+            {
+                Colors = colors;
+                WriteLine(text);
+            }
+            finally
+            {
+                Colors = _colors;
+            }
+        }
+
+        public void Write(Colors colors, string text)
+        {
+            var _colors = Colors;
+            try
+            {
+                Colors = colors;
+                Write(text);
+            }
+            finally
+            {
+                Colors = _colors;
+            }
+        }
+
         public void Write(ConsoleColor color, string text)
         {
             var foreground = ForegroundColor;